Nacta-Automation/requirements.txt
selenium
python-dotenv
requests
psutil
